requirements.txt
asgiref==3.11.1
brotli==1.2.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cloudinary==1.44.1
colorama==0.4.6
crispy-bootstrap5==0.7
dj-database-url==3.1.2
Django==4.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.0.0
django-crispy-forms==2.0
djangorestframework==3.14.0
dotenv==0.9.9
gunicorn==25.1.0
h11==0.16.0
idna==3.11
joblib==1.5.3
numpy==2.4.2
packaging==26.0
pandas==3.0.0
pillow==12.1.1
psycopg2-binary==2.9.11
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2025.2
requests==2.31.0
scikit-learn==1.8.0
scipy==1.17.0
setuptools==82.0.0
six==1.17.0
sqlparse==0.5.5
threadpoolctl==3.6.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
wheel==0.46.3
whitenoise==6.12.0
